refactor(s2n-quic-transport): refactor stream controller
